Revert back on terminal related explorer context actions registration point. Move them back to parts/execution. Dan may want to refactor it and merge with Terminal.
