+++ src/vs/workbench/parts/execution/test/electron-browser/terminalService.test.ts
@@ -17,6 +17,7 @@ suite('Execution - TerminalService', () => {
 	setup(() => {
 		mockConfig = {
 			terminal: {
+				explorerKind: 'external',
 				external: {
 					windowsExec: 'testWindowsShell',
 					osxExec: 'testOSXShell',
