+++ src/vs/workbench/parts/files/common/files.ts
@@ -61,7 +61,6 @@ export interface IFilesConfiguration extends IFilesConfiguration, IWorkbenchEdit
 		autoReveal: boolean;
 		enableDragAndDrop: boolean;
 		sortOrder: SortOrder;
-		openInTerminalKind: string;
 	};
 	editor: IEditorOptions;
 }
